separated login from DOM
