+++ src/index.js
@@ -1,12 +1,6 @@
 import { format } from 'date-fns';
 import currentLocation from './geolocation';
-import displayResults from './displayResults';
-import displayErrorMessage from './displayErrorMessage';
-
-const api = {
-  key: '2055d72a697ed8d59198d6cc43617eec',
-  baseurl: 'http://api.openweathermap.org/data/2.5/weather',
-};
+import getResults from './processWeatherData';
 
 // nodes
 
@@ -20,19 +14,6 @@ const icon = document.querySelector('.icon');
 
 const formatDate = () => format(new Date(), 'h:m a — iiii, dd MMM `yy');
 dateTime.innerHTML = formatDate();
-
-// get the value
-
-function getResults(query) {
-  fetch(`${api.baseurl}?q=${query}&appid=${api.key}`, { mode: 'cors' })
-    .then(response => response.json()).then(data => {
-      if (query === data.name) {
-        displayResults(data);
-      } else {
-        displayErrorMessage(query);
-      }
-    });
-}
 
 // set query
